feat(ai-frontend): move assistant memories into a gated sheet on the system About card
The assistant memory UI now lives as a permission-gated "Assistant Memories"
button in the About sidebar of a system's detail page (catalog SystemMetaSlot),
opening a Sheet that lists the memories saved about that specific system. This
replaces the earlier placement on the platform "About Checkstack" page, which
was the wrong surface.

- ai-frontend: add SystemMemoryButton contributing to catalog's SystemMetaSlot;
  hidden entirely (and firing no listMemories request) for users without
  ai.memory.read. Delete / always-apply stay server-enforced (ai.memory.manage).
  Remove AboutMemoriesButton; re-add the catalog-common dependency.
- about-common / about-frontend: AboutSectionsSlot stays as a general platform
  extension point (still rendered on the About page, documented generically); it
  just no longer hosts the memory button.
- Docs (ai/memory.md, frontend/extension-points.md) and bundled docs index
  regenerated; changeset added.
