+++ core/ai-frontend/src/index.tsx
@@ -3,7 +3,7 @@ import {
   createSlotExtension,
 } from "@checkstack/frontend-api";
 import { aiRoutes, pluginMetadata, aiAccess } from "@checkstack/ai-common";
-import { AboutSectionsSlot } from "@checkstack/about-common";
+import { SystemMetaSlot } from "@checkstack/catalog-common";
 import { Sparkles, Brain, Lightbulb } from "lucide-react";
 
 export default createFrontendPlugin({
@@ -49,14 +49,14 @@ export default createFrontendPlugin({
   ],
   apis: [],
   extensions: [
-    // Contribute a permission-gated "Memories" section to the platform About
-    // page. Opens a Sheet listing every memory the user can see.
-    createSlotExtension(AboutSectionsSlot, {
-      id: "ai.about.memories",
-      metadata: {},
+    // Contribute a permission-gated "Assistant Memories" button to the catalog
+    // System Details "About" sidebar. Opens a Sheet listing this system's saved
+    // memories; hidden entirely for users without ai.memory.read.
+    createSlotExtension(SystemMetaSlot, {
+      id: "ai.system-meta.memories",
       load: () =>
-        import("./components/AboutMemoriesButton").then((m) => ({
-          default: m.AboutMemoriesButton,
+        import("./components/SystemMemoryButton").then((m) => ({
+          default: m.SystemMemoryButton,
         })),
     }),
   ],
